backend/requirements.txt
Django==5.0.10
djangorestframework==3.15.2
psycopg[binary]==3.2.1
python-dotenv==1.0.1
dj-database-url==2.2.0
redis==5.0.8